apps/api/requirements.txt
# TechDados BFF - FastAPI
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# JWT / Auth
python-jose[cryptography]>=3.3.0
httpx>=0.25.0

# Logging
structlog>=24.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Optional (for future)
# redis>=5.0.0
# asyncpg>=0.29.0
# sqlalchemy>=2.0.0
